move web.py, easyargs.py, display.py, filecomp.py and dict_comp.py from pandokia/helpers/  to utils/helpers
