+++ tests/tests_new/pdkrun/all.py
@@ -4,7 +4,7 @@ import json
 
 import pandokia.helpers.test_in_db as test_in_db
 import pandokia.helpers.pycode as pycode
-import pandokia.helpers.filecomp as filecomp
+import utils.helpers.filecomp as filecomp
 
 
 f = open('test_list', 'r')
